+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Improvements
 - Improved the way InputHandler down time calculation works
 - Allow explicitly specifying each region for extended auto tiles
 - Added a generic version of IGenericDataHolder.SetData
+- Allow formatting codes to have an arbitrary custom width
 - **Drastically improved StaticSpriteBatch batching performance**
 - **Made GenericFont and TokenizedString support UTF-32 characters like emoji**
 
@@ -44,6 +45,7 @@ Fixes
 Removals
 - Removed DataContract attribute from GenericDataHolder
 - Marked EnumHelper as obsolete due to its reimplementation in [DynamicEnums](https://www.nuget.org/packages/DynamicEnums)
+- Marked Code.GetReplacementString as obsolete
 
 ### MLEM.Ui
 Additions